requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
deepface==0.0.93
numpy<2.0.0
Pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
